Refactor kernel layer util functions
