+++ pkg/kernel/mcp/iptables.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/modelcontextprotocol/go-sdk/mcp"
 	"github.com/ovn-kubernetes/ovn-kubernetes-mcp/pkg/kernel/types"
+	"github.com/ovn-kubernetes/ovn-kubernetes-mcp/pkg/utils"
 )
 
 // GetIptables MCP handler for iptables operations.
@@ -25,27 +26,34 @@ func (s *MCPServer) GetIptables(ctx context.Context, req *mcp.CallToolRequest, i
 	if err := validateIptablesCommand(in.Command); err != nil {
 		return nil, types.Result{}, fmt.Errorf("error while getting list of iptables rules: %w", err)
 	}
-	if in.FilterParameters != "" {
-		if err := validateParameters(in.FilterParameters); err != nil {
-			return nil, types.Result{}, fmt.Errorf("error while getting list of iptables rules: %w", err)
-		}
+
+	if err := utils.ValidateSafeString(in.FilterParameters, "filter parameters", true, utils.ShellMetaCharactersTypeDefault); err != nil {
+		return nil, types.Result{}, fmt.Errorf("error while getting list of iptables rules: %w", err)
 	}
 
-	cmd := newCommand(iptablesCommand(in.FilterParameters))
+	table := strings.TrimSpace(in.Table)
+	command := strings.TrimSpace(in.Command)
+	cmd := utils.NewCommand(iptablesCommand(in.FilterParameters))
 	// Defaults to 'filter' table when not specified
-	cmd.addIf(in.Table == "", "-t", "filter")
-	cmd.addIfNotEmpty(in.Table, "-t", in.Table)
+	cmd.AddIf(table == "", "-t", "filter")
+	cmd.AddIfNotEmpty(table, "-t", table)
 	// Defaults to -L (list) when command not specified
-	cmd.addIf(in.Command == "", "-L")
-	cmd.addIfNotEmpty(in.Command, in.Command)
+	cmd.AddIf(command == "", "-L")
+	cmd.AddIfNotEmpty(command, command)
 	// FilterParameters are invalid with -S/--list-rules command
-	cmd.addIf(in.FilterParameters != "" && in.Command != "-S" && in.Command != "--list-rules", strings.Fields(in.FilterParameters)...)
+	cmd.AddIf(in.FilterParameters != "" && command != "-S" && command != "--list-rules", strings.Fields(in.FilterParameters)...)
 
-	stdout, err := s.executeCommand(ctx, req, in.Node, cmd.build())
+	stdout, err := s.executeCommand(ctx, req, in.Node, cmd.Build())
 	if err != nil {
 		return nil, types.Result{}, fmt.Errorf("error while getting list of iptables rules: %w", err)
 	}
-	stdout = limitOutputLines(stdout, in.MaxLines)
+
+	// Strip empty lines from the output
+	lines := utils.StripEmptyLines(strings.Split(stdout, "\n"))
+	// Apply the head and tail parameters to the lines
+	lines = in.HeadTailParams.Apply(lines, defaultMaxOutputLines)
+	// Join the lines back into a single string
+	stdout = strings.Join(lines, "\n")
 	return nil, types.Result{Data: stdout}, nil
 }
